requirements.txt
streamlit==1.32.2
numpy==1.26.4
pillow